requirements.txt
seaborn
matplotlib
streamlit
pandas
streamlit
scikit-learn
plotly-express
bertopic
openpyxl